+++ src/core.rs
@@ -1,5 +1,19 @@
 // ## API Module
 
+use crate::sql::{
+    begin, connect, json_to_string, lock_connection, query, query_one, query_tx, query_value,
+    query_value_tx, DbConnection, JsonRow, VecInto,
+};
+use std::{fmt::Display, io::Write, path::Path as FilePath};
+
+use anyhow::Result;
+use indexmap::IndexMap;
+use minijinja::{path_loader, Environment};
+use regex::Regex;
+use serde::{Deserialize, Serialize};
+use serde_json::{from_str, json, to_value, Value as JsonValue};
+use tabwriter::TabWriter;
+
 #[derive(Debug)]
 pub enum RelatableError {
     /// An error in the configuration of a ChangeSet:
@@ -33,3 +47,918 @@ impl std::fmt::Display for RelatableError {
 }
 
 impl std::error::Error for RelatableError {}
+
+#[derive(Debug)]
+pub struct Relatable {
+    pub connection: DbConnection,
+    // pub minijinja: Environment<'static>,
+    pub default_limit: usize,
+    pub max_limit: usize,
+}
+
+impl Relatable {
+    pub async fn default() -> Result<Self> {
+        // Set up database connection.
+        let path = ".relatable/relatable.db";
+        let connection = connect(path).await?;
+        Ok(Self {
+            connection,
+            // minijinja: env,
+            default_limit: 100,
+            max_limit: 1000,
+        })
+    }
+
+    pub fn render<T: Serialize>(&self, template: &str, context: T) -> Result<String> {
+        // TODO: Optionally we should set up the environment once and store it,
+        // but during development it's very convenient to rebuild every time.
+        let mut env = Environment::new();
+
+        // Load default template strings at compile time.
+        let templates = IndexMap::from([("table.html", include_str!("templates/table.html"))]);
+
+        // Load templates dynamically if src/templates/ exists,
+        // otherwise use strings from compile time.
+        // TODO: This should be a configuration option.
+        let dir = "src/templates/";
+        if FilePath::new(dir).is_dir() {
+            env.set_loader(path_loader(dir));
+        };
+        for (name, content) in templates {
+            match env.get_template(name) {
+                Ok(_) => (),
+                Err(_) => env.add_template(name, content).unwrap(),
+            }
+        }
+
+        env.get_template(template)?
+            .render(context)
+            .map_err(|e| e.into())
+    }
+
+    pub async fn get_table(&self, table_name: &str) -> Result<Table> {
+        let statement = r#"SELECT max(change_id) FROM history WHERE "table" = ?"#;
+        let params = json!([table_name]);
+        let change_id = match query_value(&self.connection, &statement, Some(&params)).await? {
+            Some(value) => value.as_u64().unwrap_or_default() as usize,
+            None => 0,
+        };
+        Ok(Table {
+            name: table_name.to_string(),
+            change_id,
+            ..Default::default()
+        })
+    }
+
+    pub fn from(&self, table_name: &str) -> Select {
+        Select {
+            table_name: table_name.to_string(),
+            limit: self.default_limit,
+            ..Default::default()
+        }
+    }
+
+    pub async fn fetch_columns(&self, table_name: &str) -> Result<Vec<Column>> {
+        // WARN: SQLite only!
+        let statement = format!(r#"SELECT * FROM pragma_table_info("{table_name}");"#);
+        Ok(query(&self.connection, &statement, None)
+            .await?
+            .iter()
+            .map(|row| Column {
+                name: row.get_string("name"),
+            })
+            .filter(|c| !c.name.starts_with("_"))
+            .collect())
+    }
+
+    pub async fn fetch(&self, select: &Select) -> Result<ResultSet> {
+        let table = self.get_table(&select.table_name).await?;
+        let columns = self.fetch_columns(&select.table_name).await?;
+        let (statement, params) = select.to_sqlite()?;
+        tracing::debug!("SQL {statement}");
+        let params = json!(params);
+        let json_rows = query(&self.connection, &statement, Some(&params)).await?;
+
+        let count = json_rows.len();
+        let total = match json_rows.get(0) {
+            Some(row) => row
+                .content
+                .get("_total")
+                .and_then(|x| x.as_u64())
+                .unwrap_or(0) as usize,
+            None => 0,
+        };
+
+        let rows: Vec<Row> = json_rows.vec_into();
+        Ok(ResultSet {
+            range: Range {
+                count,
+                total,
+                start: select.offset + 1,
+                end: select.offset + count,
+            },
+            select: select.clone(),
+            table,
+            columns,
+            rows,
+        })
+    }
+
+    pub async fn fetch_json_rows(&self, select: &Select) -> Result<Vec<JsonRow>> {
+        let (statement, params) = select.to_sqlite()?;
+        let params = json!(params);
+        query(&self.connection, &statement, Some(&params)).await
+    }
+
+    pub async fn set_values(&self, changeset: &ChangeSet) -> Result<()> {
+        let action = changeset.action.to_string();
+        let user = changeset.user.clone();
+        let table = changeset.table.clone();
+        let description = changeset.description.clone();
+
+        // Get the connection and begin a transaction:
+        let mut locked_conn = lock_connection(&self.connection).await;
+        let mut tx = begin(&self.connection, &mut locked_conn).await?;
+
+        // Make sure the user is present.
+        let color = random_color::RandomColor::new().to_hex();
+        let statement = r#"INSERT OR IGNORE INTO user("name", "color") VALUES (?, ?)"#;
+        let params = json!([user, color]);
+        query_tx(&mut tx, &statement, Some(&params)).await?;
+
+        // Update the user's cursor position.
+        let cursor = changeset.to_cursor()?;
+        let statement =
+            r#"UPDATE user SET "cursor" = ?, "datetime" = CURRENT_TIMESTAMP WHERE "name" = ?"#;
+        let params = json!([to_value(cursor).unwrap_or_default(), user]);
+        query_value_tx(&mut tx, &statement, Some(&params)).await?;
+
+        let statement = r#"INSERT INTO change('user', 'action', 'table', 'description', 'content')
+                           VALUES (?, ?, ?, ?, ?)
+                           RETURNING change_id"#;
+        let content = to_value(&changeset.changes).unwrap_or_default();
+        let params = json!([user, action, table, description, content]);
+        let change_id = query_value_tx(&mut tx, &statement, Some(&params)).await?;
+        let change_id = change_id
+            .expect("a change_id")
+            .as_u64()
+            .expect("an integer");
+
+        for change in &changeset.changes {
+            tracing::debug!("CHANGE {change:?}");
+            match change {
+                Change::Update { row, column, value } => {
+                    let statement = r#"INSERT INTO history
+                                       ('change_id', 'table', 'row', 'before', 'after')
+                                       VALUES (?, ?, ?, 'TODO', 'TODO')
+                                       RETURNING history_id"#;
+                    let params = json!([change_id, table, row]);
+                    query_value_tx(&mut tx, &statement, Some(&params)).await?;
+
+                    // WARN: This just sets text!
+                    let statement =
+                        format!(r#"UPDATE "{table}" SET "{column}" = ? WHERE _id = ?"#,);
+                    // TODO: Render JSON to SQL properly.
+                    let value = json_to_string(value);
+                    let params = json!([value, row]);
+                    query_tx(&mut tx, &statement, Some(&params)).await?;
+                }
+            }
+        }
+
+        // Commit the transaction:
+        tx.commit().await?;
+
+        Ok(())
+    }
+
+    pub async fn get_user(&self, username: &str) -> Account {
+        let statement = format!(r#"SELECT * FROM user WHERE name = '{username}' LIMIT 1"#);
+        let user = query_one(&self.connection, &statement).await;
+        if let Ok(user) = user {
+            if let Some(user) = user {
+                return Account {
+                    name: username.to_string(),
+                    color: user.get_string("color"),
+                };
+            }
+        }
+        Account {
+            ..Default::default()
+        }
+    }
+
+    pub async fn get_users(&self) -> Result<IndexMap<String, UserCursor>> {
+        let mut users = IndexMap::new();
+        // let statement = format!(
+        //     r#"SELECT * FROM user WHERE cursor IS NOT NULL
+        //        AND "datetime" >= DATETIME('now', '-10 minutes')"#
+        // );
+        let statement = format!(r#"SELECT * FROM user WHERE cursor IS NOT NULL"#);
+        let rows = query(&self.connection, &statement, None).await?;
+        for row in rows {
+            let name = row.get_string("name");
+            users.insert(
+                name.clone(),
+                UserCursor {
+                    name: name.clone(),
+                    color: row.get_string("color"),
+                    cursor: from_str(&row.get_string("cursor"))?,
+                    datetime: row.get_string("datetime"),
+                },
+            );
+        }
+        Ok(users)
+    }
+
+    pub async fn get_tables(&self) -> Result<IndexMap<String, Table>> {
+        let mut tables = IndexMap::new();
+        let statement = format!(
+            r#"SELECT *,
+                 (SELECT max(change_id)
+                  FROM history
+                  WHERE history."table" = "table"."table"
+                 ) AS _change_id
+               FROM 'table'"#
+        );
+
+        let rows = query(&self.connection, &statement, None).await?;
+        for row in rows {
+            let name = row.get_string("table");
+            tables.insert(
+                name.clone(),
+                Table {
+                    name: name.clone(),
+                    change_id: row
+                        .content
+                        .get("_change_id")
+                        .and_then(|i| i.as_u64())
+                        .unwrap_or_default() as usize,
+                    ..Default::default()
+                },
+            );
+        }
+        Ok(tables)
+    }
+
+    pub async fn get_site(&self, username: &str) -> Site {
+        let mut users = self.get_users().await.unwrap_or_default();
+        users.shift_remove(username);
+        Site {
+            title: "RLTBL".to_string(),
+            root: "".to_string(),
+            user: self.get_user(username).await,
+            users,
+            tables: self.get_tables().await.unwrap_or_default(),
+        }
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct ChangeSet {
+    pub action: ChangeAction,
+    pub table: String,
+    pub user: String,
+    pub description: String,
+    pub changes: Vec<Change>,
+}
+
+impl ChangeSet {
+    fn to_cursor(&self) -> Result<Cursor> {
+        let table = self.table.clone();
+        match self.changes.first() {
+            Some(change) => match change {
+                Change::Update {
+                    row,
+                    column,
+                    value: _,
+                } => Ok(Cursor {
+                    table,
+                    row: *row,
+                    column: column.to_string(),
+                }),
+            },
+            None => Err(RelatableError::ChangeError("No changes in set".into()).into()),
+        }
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub enum ChangeAction {
+    Do,
+    Undo,
+    Redo,
+}
+
+impl Display for ChangeAction {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            ChangeAction::Do => write!(f, "do"),
+            ChangeAction::Undo => write!(f, "undo"),
+            ChangeAction::Redo => write!(f, "redo"),
+        }
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+#[serde(tag = "type")]
+pub enum Change {
+    Update {
+        row: usize,
+        column: String,
+        value: JsonValue,
+    },
+    // Add
+    // Delete
+    // Move
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct Cell {
+    value: JsonValue,
+    text: String,
+}
+
+impl From<&JsonValue> for Cell {
+    fn from(value: &JsonValue) -> Self {
+        Self {
+            value: value.clone(),
+            text: match value {
+                JsonValue::String(value) => value.to_string(),
+                value => format!("{value}"),
+            },
+        }
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct Row {
+    id: usize,
+    order: usize,
+    change_id: usize,
+    cells: IndexMap<String, Cell>,
+}
+
+impl Row {
+    fn to_strings(&self) -> Vec<String> {
+        self.cells.values().map(|cell| cell.text.clone()).collect()
+    }
+}
+
+// Ignore columns that start with "_"
+impl From<JsonRow> for Row {
+    fn from(row: JsonRow) -> Self {
+        Self {
+            id: row
+                .content
+                .get("_id")
+                .and_then(|i| i.as_u64())
+                .unwrap_or_default() as usize,
+            order: row
+                .content
+                .get("_order")
+                .and_then(|i| i.as_u64())
+                .unwrap_or_default() as usize,
+            change_id: row
+                .content
+                .get("_change_id")
+                .and_then(|i| i.as_u64())
+                .unwrap_or_default() as usize,
+            cells: row
+                .content
+                .iter()
+                .filter(|(k, _)| !k.starts_with("_"))
+                .map(|(k, v)| (k.clone(), v.into()))
+                .collect(),
+        }
+    }
+}
+
+impl From<Row> for Vec<String> {
+    fn from(row: Row) -> Self {
+        row.to_strings()
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct Column {
+    name: String,
+    // sqltype: String,
+}
+
+#[derive(Clone, Debug, Default, Serialize, Deserialize)]
+pub struct Table {
+    name: String,
+    // The history_id of the most recent update to this table.
+    change_id: usize,
+    editable: bool,
+}
+
+#[derive(Clone, Debug, Default, Serialize, Deserialize)]
+pub struct Range {
+    count: usize,
+    total: usize,
+    start: usize,
+    end: usize,
+}
+
+impl std::fmt::Display for Range {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "Rows {}-{} of {}", self.start, self.end, self.total)
+    }
+}
+
+#[derive(Clone, Debug, Default, Serialize, Deserialize)]
+pub struct ResultSet {
+    select: Select,
+    range: Range,
+    table: Table,
+    columns: Vec<Column>,
+    rows: Vec<Row>,
+}
+
+impl std::fmt::Display for ResultSet {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let mut tw = TabWriter::new(vec![]);
+        tw.write(format!("{}\n", self.range).as_bytes())
+            .unwrap_or_default();
+        let header = &self
+            .columns
+            .iter()
+            .map(|c| c.name.clone())
+            .collect::<Vec<String>>();
+        tw.write(format!("{}\n", header.join("\t")).as_bytes())
+            .unwrap_or_default();
+        for row in &self.rows {
+            tw.write(format!("{}\n", row.to_strings().join("\t")).as_bytes())
+                .unwrap_or_default();
+        }
+        tw.flush().expect("TabWriter to flush");
+        let written = String::from_utf8(tw.into_inner().unwrap()).unwrap();
+        write!(f, "{written}")
+    }
+}
+
+pub type QueryParams = IndexMap<String, String>;
+
+pub enum Format {
+    Html,
+    Json,
+    PrettyJson,
+    Default,
+}
+
+impl TryFrom<&String> for Format {
+    fn try_from(path: &String) -> Result<Self> {
+        let path = path.to_lowercase();
+        let format = if path.ends_with(".pretty.json") {
+            Format::PrettyJson
+        } else if path.ends_with(".json") {
+            Format::Json
+        } else if path.ends_with(".html") || path.ends_with(".htm") {
+            Format::Html
+        } else if path.contains(".") {
+            return Err(
+                RelatableError::FormatError(format!("Unknown format for path {path}")).into(),
+            );
+        } else {
+            Format::Default
+        };
+        Ok(format)
+    }
+
+    type Error = anyhow::Error;
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub enum Filter {
+    Equal { column: String, value: JsonValue },
+    GreaterThan { column: String, value: JsonValue },
+    GreaterThanOrEqual { column: String, value: JsonValue },
+    LessThan { column: String, value: JsonValue },
+    LessThanOrEqual { column: String, value: JsonValue },
+    Is { column: String, value: JsonValue },
+    In { column: String, value: JsonValue },
+}
+
+fn render_in_not_in<S: Into<String>>(
+    lhs: S,
+    options: &Vec<JsonValue>,
+    positive: bool,
+) -> Result<String> {
+    let negation;
+    if !positive {
+        negation = " NOT";
+    } else {
+        negation = "";
+    }
+
+    let mut values = vec![];
+    let mut is_string_list = false;
+    for (i, option) in options.iter().enumerate() {
+        match option {
+            JsonValue::String(s) => {
+                if i == 0 {
+                    is_string_list = true;
+                } else if !is_string_list {
+                    return Err(RelatableError::InputError(format!(
+                        "{:?} contains both text and numeric types.",
+                        options
+                    ))
+                    .into());
+                }
+                values.push(format!("{s}"))
+            }
+            JsonValue::Number(n) => {
+                if i == 0 {
+                    is_string_list = false;
+                } else if is_string_list {
+                    return Err(RelatableError::InputError(format!(
+                        "{:?} contains both text and numeric types.",
+                        options
+                    ))
+                    .into());
+                }
+                values.push(format!("{n}"))
+            }
+            _ => {
+                return Err(RelatableError::InputError(format!(
+                    "{:?} is not an array of strings or numbers.",
+                    options
+                ))
+                .into())
+            }
+        };
+    }
+    let value_list = format!("({})", values.join(", "));
+    let filter_sql = format!("{}{} IN {}", lhs.into(), negation, value_list);
+    Ok(filter_sql)
+}
+
+impl Display for Filter {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        // TODO: This should be factored out.
+        fn json_to_string(value: &JsonValue) -> String {
+            match value {
+                JsonValue::Null => "NULL".to_string(),
+                JsonValue::Bool(value) => value.to_string(),
+                JsonValue::Number(value) => value.to_string(),
+                JsonValue::String(value) => format!("'{value}'"),
+                JsonValue::Array(value) => format!("'{value:?}'"),
+                JsonValue::Object(value) => format!("'{value:?}'"),
+            }
+        }
+        let result = match self {
+            Filter::Equal { column, value } => {
+                let value = json_to_string(&value);
+                format!(r#""{column}" = {value}"#)
+            }
+            Filter::GreaterThan { column, value } => {
+                let value = json_to_string(&value);
+                format!(r#""{column}" > {value}"#)
+            }
+            Filter::GreaterThanOrEqual { column, value } => {
+                let value = json_to_string(&value);
+                format!(r#""{column}" >= {value}"#)
+            }
+            Filter::LessThan { column, value } => {
+                let value = json_to_string(&value);
+                format!(r#""{column}" < {value}"#)
+            }
+            Filter::LessThanOrEqual { column, value } => {
+                let value = json_to_string(&value);
+                format!(r#""{column}" <= {value}"#)
+            }
+            Filter::Is { column, value } => {
+                // Note that we are presupposing SQLite syntax which is not universal for IS:
+                let value = json_to_string(&value);
+                format!(r#""{column}" IS {value}"#)
+            }
+            Filter::In { column, value } => {
+                if let JsonValue::Array(values) = value {
+                    let filter_str = match render_in_not_in(column, values, true) {
+                        Err(_) => return Err(std::fmt::Error),
+                        Ok(filter_str) => filter_str,
+                    };
+                    format!("{filter_str}")
+                } else {
+                    return Err(std::fmt::Error);
+                }
+            }
+        };
+        write!(f, "{result}")
+    }
+}
+
+#[derive(Clone, Debug, Default, Serialize, Deserialize)]
+pub struct Select {
+    pub table_name: String,
+    pub limit: usize,
+    pub offset: usize,
+    pub filters: Vec<Filter>,
+}
+
+impl Select {
+    pub fn from_path_and_query(rltbl: &Relatable, path: &str, query_params: &QueryParams) -> Self {
+        let table_name = path.split(".").next().unwrap_or_default().to_string();
+        let mut filters = Vec::new();
+        for (column, pattern) in query_params {
+            if pattern.starts_with("eq.") {
+                let column = column.to_string();
+                let value = serde_json::from_str(&pattern.replace("eq.", ""));
+                match value {
+                    Ok(value) => filters.push(Filter::Equal { column, value }),
+                    Err(_) => tracing::warn!("invalid filter value {pattern}"),
+                }
+            } else if pattern.starts_with("gt.") {
+                let column = column.to_string();
+                let value = serde_json::from_str(&pattern.replace("gt.", ""));
+                match value {
+                    Ok(value) => filters.push(Filter::GreaterThan { column, value }),
+                    Err(_) => tracing::warn!("invalid filter value {pattern}"),
+                }
+            } else if pattern.starts_with("gte.") {
+                let column = column.to_string();
+                let value = serde_json::from_str(&pattern.replace("gte.", ""));
+                match value {
+                    Ok(value) => filters.push(Filter::GreaterThanOrEqual { column, value }),
+                    Err(_) => tracing::warn!("invalid filter value {pattern}"),
+                }
+            } else if pattern.starts_with("lt.") {
+                let column = column.to_string();
+                let value = serde_json::from_str(&pattern.replace("lt.", ""));
+                match value {
+                    Ok(value) => filters.push(Filter::LessThan { column, value }),
+                    Err(_) => tracing::warn!("invalid filter value {pattern}"),
+                }
+            } else if pattern.starts_with("lte.") {
+                let column = column.to_string();
+                let value = serde_json::from_str(&pattern.replace("lte.", ""));
+                match value {
+                    Ok(value) => filters.push(Filter::LessThanOrEqual { column, value }),
+                    Err(_) => tracing::warn!("invalid filter value {pattern}"),
+                }
+            } else if pattern.starts_with("is.") {
+                let column = column.to_string();
+                let value = pattern.replace("is.", "");
+                match value.to_lowercase().as_str() {
+                    "null" => filters.push(Filter::Is {
+                        column,
+                        value: JsonValue::Null,
+                    }),
+                    _ => match serde_json::from_str(&value) {
+                        Ok(value) => filters.push(Filter::Is { column, value }),
+                        Err(_) => tracing::warn!("invalid filter value {pattern}"),
+                    },
+                };
+            } else if pattern.starts_with("in.") {
+                let column = column.to_string();
+                let value = pattern.replace("in.", "");
+                println!("COLUMN: {column}, VALUE: {value}");
+                todo!()
+            }
+        }
+        let limit: usize = query_params
+            .get("limit")
+            .and_then(|x| x.parse::<usize>().ok())
+            .unwrap_or(rltbl.default_limit)
+            .min(rltbl.max_limit);
+        let offset: usize = query_params
+            .get("offset")
+            .and_then(|x| x.parse::<usize>().ok())
+            .unwrap_or_default();
+        Self {
+            table_name,
+            limit,
+            offset,
+            filters,
+            ..Default::default()
+        }
+    }
+
+    pub fn limit(mut self, limit: &usize) -> Self {
+        self.limit = *limit;
+        self
+    }
+
+    pub fn offset(mut self, offset: &usize) -> Self {
+        self.offset = *offset;
+        self
+    }
+
+    pub fn filters(mut self, filters: &Vec<String>) -> Result<Self> {
+        let eq = Regex::new(r#"^(\w+)\s*=\s*"?(\w+)"?$"#).unwrap();
+        let gt = Regex::new(r"^(\w+)\s*>\s*(\w+)$").unwrap();
+        let gte = Regex::new(r"^(\w+)\s*>=\s*(\w+)$").unwrap();
+        let lt = Regex::new(r"^(\w+)\s*<\s*(\w+)$").unwrap();
+        let lte = Regex::new(r"^(\w+)\s*<=\s*(\w+)$").unwrap();
+        let is = Regex::new(r#"^(\w+)\s+(IS|is)\s+"?(\w+)"?$"#).unwrap();
+        let is_in = Regex::new(r#"^(\w+)\s+(IN|in)\s+\((\w+(,\s*\w+)*)\)$"#).unwrap();
+        // Used for text types:
+        let maybe_quote_value = |value: &str| -> Result<JsonValue> {
+            if value.starts_with("\"") {
+                let value = serde_json::from_str(&value)?;
+                Ok(value)
+            } else {
+                let value = serde_json::from_str(&format!(r#""{value}""#))?;
+                Ok(value)
+            }
+        };
+        for filter in filters {
+            if eq.is_match(&filter) {
+                let captures = eq.captures(&filter).unwrap();
+                let column = captures.get(1).unwrap().as_str().to_string();
+                let value = &captures.get(2).unwrap().as_str();
+                let value = maybe_quote_value(&value)?;
+                self.filters.push(Filter::Equal { column, value });
+            } else if gt.is_match(&filter) {
+                let captures = gt.captures(&filter).unwrap();
+                let column = captures.get(1).unwrap().as_str().to_string();
+                let value = &captures.get(2).unwrap().as_str();
+                let value = serde_json::from_str(&value)?;
+                self.filters.push(Filter::GreaterThan { column, value });
+            } else if gte.is_match(&filter) {
+                let captures = gte.captures(&filter).unwrap();
+                let column = captures.get(1).unwrap().as_str().to_string();
+                let value = &captures.get(2).unwrap().as_str();
+                let value = serde_json::from_str(&value)?;
+                self.filters
+                    .push(Filter::GreaterThanOrEqual { column, value });
+            } else if lt.is_match(&filter) {
+                let captures = lt.captures(&filter).unwrap();
+                let column = captures.get(1).unwrap().as_str().to_string();
+                let value = &captures.get(2).unwrap().as_str();
+                let value = serde_json::from_str(&value)?;
+                self.filters.push(Filter::LessThan { column, value });
+            } else if lte.is_match(&filter) {
+                let captures = lte.captures(&filter).unwrap();
+                let column = captures.get(1).unwrap().as_str().to_string();
+                let value = &captures.get(2).unwrap().as_str();
+                let value = serde_json::from_str(&value)?;
+                self.filters.push(Filter::LessThanOrEqual { column, value });
+            } else if is.is_match(&filter) {
+                let captures = is.captures(&filter).unwrap();
+                let column = captures.get(1).unwrap().as_str().to_string();
+                let value = &captures.get(3).unwrap().as_str();
+                let value = match value.to_lowercase().as_str() {
+                    "null" => JsonValue::Null,
+                    _ => maybe_quote_value(&value)?,
+                };
+                self.filters.push(Filter::Is { column, value });
+            } else if is_in.is_match(&filter) {
+                let captures = is_in.captures(&filter).unwrap();
+                let column = captures.get(1).unwrap().as_str().to_string();
+                let values = &captures.get(3).unwrap().as_str();
+                let separator = Regex::new(r"\s*,\s*").unwrap();
+                let values = separator
+                    .split(values)
+                    .map(|v| serde_json::from_str::<JsonValue>(v).unwrap_or(json!(v.to_string())))
+                    .collect::<Vec<_>>();
+                self.filters.push(Filter::In {
+                    column,
+                    value: json!(values),
+                });
+            } else {
+                return Err(RelatableError::ConfigError(format!("invalid filter {filter}")).into());
+            }
+        }
+        Ok(self)
+    }
+
+    pub fn eq<T>(mut self, column: &str, value: &T) -> Result<Self>
+    where
+        T: Serialize,
+    {
+        self.filters.push(Filter::Equal {
+            column: column.to_string(),
+            value: to_value(value)?,
+        });
+        Ok(self)
+    }
+
+    pub fn gt<T>(mut self, column: &str, value: &T) -> Result<Self>
+    where
+        T: Serialize,
+    {
+        self.filters.push(Filter::GreaterThan {
+            column: column.to_string(),
+            value: to_value(value)?,
+        });
+        Ok(self)
+    }
+
+    pub fn gte<T>(mut self, column: &str, value: &T) -> Result<Self>
+    where
+        T: Serialize,
+    {
+        self.filters.push(Filter::GreaterThanOrEqual {
+            column: column.to_string(),
+            value: to_value(value)?,
+        });
+        Ok(self)
+    }
+
+    pub fn lt<T>(mut self, column: &str, value: &T) -> Result<Self>
+    where
+        T: Serialize,
+    {
+        self.filters.push(Filter::LessThan {
+            column: column.to_string(),
+            value: to_value(value)?,
+        });
+        Ok(self)
+    }
+
+    pub fn lte<T>(mut self, column: &str, value: &T) -> Result<Self>
+    where
+        T: Serialize,
+    {
+        self.filters.push(Filter::LessThanOrEqual {
+            column: column.to_string(),
+            value: to_value(value)?,
+        });
+        Ok(self)
+    }
+
+    pub fn is<T>(mut self, column: &str, value: &T) -> Result<Self>
+    where
+        T: Serialize,
+    {
+        self.filters.push(Filter::Is {
+            column: column.to_string(),
+            value: to_value(value)?,
+        });
+        Ok(self)
+    }
+
+    pub fn is_in<T>(mut self, column: &str, value: &T) -> Result<Self>
+    where
+        T: Serialize,
+    {
+        self.filters.push(Filter::In {
+            column: column.to_string(),
+            value: to_value(value)?,
+        });
+        Ok(self)
+    }
+
+    pub fn to_sqlite(&self) -> Result<(String, Vec<JsonValue>)> {
+        tracing::debug!("to_sqlite: {self:?}");
+        let table = &self.table_name;
+        let mut lines = Vec::new();
+        lines.push("SELECT *,".to_string());
+        // WARN: The _total count should probably be optional.
+        lines.push("  COUNT(1) OVER() AS _total,".to_string());
+        lines.push(format!(
+            r#"  (SELECT MAX(change_id) FROM history
+                   WHERE "table" = ?
+                     AND "row" = _id
+                 ) AS _change_id"#
+        ));
+        lines.push(format!(r#"FROM "{}""#, self.table_name));
+        for (i, filter) in self.filters.iter().enumerate() {
+            let keyword = if i == 0 { "WHERE" } else { "  AND" };
+            lines.push(format!("{keyword} {filter}"));
+        }
+        if self.limit > 0 {
+            lines.push(format!("LIMIT {}", self.limit));
+        }
+        if self.offset > 0 {
+            lines.push(format!("OFFSET {}", self.offset));
+        }
+
+        Ok((lines.join("\n"), vec![json!(table)]))
+    }
+}
+
+// Web Site Stuff
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct Site {
+    title: String,
+    root: String,
+    user: Account,
+    users: IndexMap<String, UserCursor>,
+    tables: IndexMap<String, Table>,
+}
+
+#[derive(Clone, Debug, Default, Serialize, Deserialize)]
+pub struct Account {
+    name: String,
+    color: String,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct Cursor {
+    table: String,
+    row: usize,
+    column: String,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct UserCursor {
+    name: String,
+    color: String,
+    cursor: Cursor,
+    datetime: String,
+}
