+++ src/lib.rs
@@ -2,7 +2,9 @@
 //!
 //! This is rltbl.
 
-// Sub-modules:
+///////////////////////////////////////////////////////////////////////////////
+// Sub-modules
+///////////////////////////////////////////////////////////////////////////////
 
 /// Core functionality
 pub mod core;
@@ -10,4 +12,6 @@ pub mod core;
 /// An abstraction over SQL engines
 pub mod sql;
 
-// Global constants and other lookups:
+///////////////////////////////////////////////////////////////////////////////
+// Global constants and other lookups
+///////////////////////////////////////////////////////////////////////////////
